Recreate designate_pdns tables (bsc#1085618)
DB schema for Designate backend driver for PowerDNS v4+ has changed.
This change is not handled by DB migration scripts.

In the previous implementation we were dropping the DB and recreating it.
That lead to permission issues for pdns_server. To overcome this, now
we discover if we have tables from old DB schema, If so, we drop them
and recreate them based on new schema.

Also a slip in webserver-allow-from config option is corrected.

Change-Id: Ic99958d0c535e86511187ab4fd9067a9da36a7ab