PowerDNS deploy fails on mid scale (bsc#1094087)
PowerDNS post configure play fails on midscale input model.
This is due to wrong hosts selection for the play.
In the current state of things for PowerDNS, we need to run SQL
commands to recreate the powerdns DB. We do this in powerdns
post configure play. The correct hosts for the play thus is
FND-MDB, not DES-PMG or DES-WKR.

This was working for models where DB and PowerDNS were co located.

This patch corrects the hosts selection for the play.

Change-Id: I03d69ad03e4837bf6e5a629402c32c7dd537b496